almost finished simplifying chat tree
