--- screens/main/chat/components/Input/styles.js
@@ -1,23 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-import { COLORS } from '../../styles'
-
-export default StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-around',
-    width: '100%'
-  },
-  inputContainer: {
-    width: '70%'
-  },
-  input: {
-    height: 40,
-    borderColor: COLORS.GREY,
-    borderWidth: 1,
-    borderRadius: 3,
-    flexDirection: 'row',
-    paddingHorizontal: 10
-  }
-})
