--- screens/main/chat/components/common/Button/styles.js
@@ -1,16 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-import { COLORS } from '../../../styles'
-
-export default StyleSheet.create({
-  container: {
-    marginHorizontal: 20,
-    paddingVertical: 10,
-    paddingHorizontal: 16,
-    backgroundColor: COLORS.PRIMARY,
-    borderRadius: 3
-  },
-  text: {
-    color: COLORS.WHITE
-  }
-})
